cube: Enable console output on Windows
If vkcube is run from a command prompt, cube will now discover the console and
write to it.

The main thin to notice is how there is a call to SendMessage right before cube
exits. This is intentionally because without it, when cube is launched from
command prompt, the console doesn't realize that cube has exited and will wait
for a final keystroke before returning to normal. By having cube send that
keystroke, this forces the console to return to normal.